Move normaliseContent to doc
